Revert "Remove ConfigParser class and move generateMap method to ConfigUtils."
This reverts commit d3b3084f31948d2f61497a62199bf86955aaabd4.

# Conflicts:
#	src/Optimizely/Utils/ConfigUtils.php
